# SCA hardening for modular field inversion in `EC_GROUP`s