… crypto/ec/ec_err.c
… include/openssl/ecerr.h