# SCA hardening for modular field inversion in `EC_GROUP`s

This PR adds a dedicated function in `EC_METHOD` to access a modular field inversion implementation suitable for the specifics of the implemented curve, featuring SCA countermeasures.

The new pointer is defined as:
```c
int (*field_inv)(const EC_GROUP *, BIGNUM *r, const BIGNUM *a, BN_CTX *)
```
and computes the multiplicative inverse of `a` in the underlying field, storing the result in `r`.

Three implementations are included, each including specific SCA countermeasures:
  - `ec_GFp_simple_field_inv()`, featuring SCA hardening through blinding.
  - `ec_GFp_mont_field_inv()`, featuring SCA hardening through *Fermat's Little Theorem* (FLT) inversion.
  - `ec_GF2m_simple_field_inv()`, that uses `BN_GF2m_mod_inv()` which already features SCA hardening through blinding.

This PR also adds a new error reason code (i.e., `EC_R_CANNOT_INVERT`) to improve consistency between the three implementations as all of them could fail for the same reason but through different code paths resulting in inconsistent error stack states.

## Security analysis

Implementing this countermeasures also helps addressing a leakage affecting conversions from projective to affine coordinates in `EC_POINT_get_affine_coordinates()` which did not consider the `BN_FLG_CONSTTIME` flag, hitting a leaky code path for inversion.

Public disclosure of this minor defect has been responsibly coordinated with the OpenSSL security team.

## Backports

- This should go to `master` and `1.1.1-stable`.
- Ideally this could be backported to `1.1.0-stable` to minimize discrepancies between it and the LTS branch and minimize potential future security fixes, but technically I am not sure if this qualifies as security fix.
- Backport to 1.0.2 is not recommended, as it would require modifying non-opaque `struct`s.


## Checklist

- [x] documentation is added
- [x] tests are added