@mattcaswell 2019-02-15
I personally think this should go to 1.1.0 as well - although I suspect it may not cherry-pick cleanly due to the tests. Anyway I'll add the label for now.