@mattcaswell 2019-02-15
Travis red cross is relevant. 

````
test/ec_internal_test.c:48:84: error: invalid operands to binary expression ('void' and 'int')
        || !test_true("test/ec_internal_test.c", 48, "BN_zero(a)", (BN_zero_ex(a)) != 0)
                                                                   ~~~~~~~~~~~~~~~ ^  ~
1 error generated.
make[1]: *** [test/ec_internal_test-bin-ec_internal_test.o] Error 1
make[1]: Leaving directory `/home/travis/build/openssl/openssl'
make: *** [build_programs] Error 2
````

That's quite interesting. `BN_zero` is a macro defined like this:
````
# if OPENSSL_API_0_9_8
#  define BN_zero(a)      BN_zero_ex(a)
# else
#  define BN_zero(a)      (BN_set_word((a),0))
# endif
````
So, depending on you build BN_zero() is either a void return or an int return!!!! That's not great. :-(