@romen 2019-02-15
> So, depending on you build BN_zero() is either a void return or an int return!!!! That's not great. :-(

Yep! I cringed when I noticed `BN_zero()` and `BN_one()` have different signatures!

Sidenote: is 3.0.0 a good candidate to make `BN_zero()` "normal"?