@mattcaswell 2019-02-15
> Sidenote: is 3.0.0 a good candidate to make `BN_zero()` "normal"?

TBH the BN_zero thing is a mess and I'm not sure what to do about it. AFAICT the default is you get:

````
#  define BN_zero(a)      (BN_set_word((a),0))
````

But if you build with no-deprecated you get a different definition for it with a different (void) return value:

````
#  define BN_zero(a)      BN_zero_ex(a)
````

I'm tempted to get rid of the no-deprecated form and revert back to the default (deprecated) form in all cases. At least that way you get a signature that is consistent with BN_one() and won't cause unneccessary breakage when we (at some point in the future) start removing deprecated stuff.

(BTW: Do you have a response to my scatter-gather comment above?)