@romen 2019-02-15
Travis red cross now is unrelated: there was some fetch issue on the osx builds.
Can we kick them to run again somehow?

> Do you have a view on the likely performance implications of this?

I will conjure some numbers, but I need more time to convince `speed` to hit all the relevant code paths.