@bbbrumley 2019-02-15
> Do you have a view on the likely performance implications of this?

Should be close to zero. See Sec. 7.2 [here](https://eprint.iacr.org/2018/651). It's in the ECC field not the group order, but similar sizes.