@romen 2019-02-15
> > Do you have a view on the likely performance implications of this?
> 
> Should be close to zero. See Sec. 7.2 [here](https://eprint.iacr.org/2018/651). It's in the ECC field not the group order, but similar sizes.

I can confirm with fresh data!

# Benchmark

- `nistp256`, using the `nistz256` implementation, is unaffected as it never hits any `ec_GF*_field_inv()`
- `nistp521`, representative of the `ec_nistp_64_gcc_128` implementation family, is unaffected as it never hits any `ec_GF*_field_inv()`
- `nistb571` uses the `EC_GF2m_simple_method`, hits the `ec_GF2m_simple_field_inv()` method, which is just a wrapper around the previous implementation which was hardened in #6070 (0dae8bafceabc8966383aa1f11ee8622f7dbde2f), so no real change in performance
- `brainpoolP512r1` uses the `EC_GFp_mont_method` and hits the `ec_GFp_mont_field_inv()` method with the FLT hardening: the comparison shows that the performance hit is insignificant
- hitting `ec_GFp_simple_field_inv()` from a named curve within the library is difficult: one way is using `nistp{192,224,256,384,521}` in builds with `no-asm`, which end up using the `EC_GFp_nist_method` which "inherits" the `ec_GFp_simple_field_inv()` function.
   Even in this case the performance backlash is minimal.

## Before this PR (fa1f03061037cbdac5369849a885c1191a2550d9)

```
> openssl speed ecdhb571 ecdhbrp512r1 ecdhp521 ecdhp256
version: 3.0.0-dev
built on: built on: Fri Feb 15 19:15:35 2019 UTC
options:bn(64,64) rc4(16x,int) des(int) aes(partial) idea(int) blowfish(ptr) 
compiler: ccache gcc -fPIC -pthread -m64 -Wa,--noexecstack -Wall -O3 -g -DOPENSSL_USE_NODELETE -DL_ENDIAN -DOPENSSL_PIC -DOPENSSL_CPUID_OBJ -DOPENSSL_IA32_SSE2 -DOPENSSL_BN_ASM_MONT -DOPENSSL_BN_ASM_MONT5 -DOPENSSL_BN_ASM_GF2m -DSHA1_ASM -DSHA256_ASM -DSHA512_ASM -DKECCAK1600_ASM -DRC4_ASM -DMD5_ASM -DAES_ASM -DVPAES_ASM -DBSAES_ASM -DGHASH_ASM -DECP_NISTZ256_ASM -DX25519_ASM -DPADLOCK_ASM -DPOLY1305_ASM -DNDEBUG
                              op      op/s
 256 bits ecdh (nistp256)   0.0001s  19587.7
 521 bits ecdh (nistp521)   0.0003s   2965.6
 571 bits ecdh (nistb571)   0.0019s    530.0
 512 bits ecdh (brainpoolP512r1)   0.0014s    715.8

> openssl speed ecdhp521 # no-asm build
version: 3.0.0-dev
built on: built on: Fri Feb 15 19:58:37 2019 UTC
options:bn(64,64) rc4(int) des(int) aes(partial) idea(int) blowfish(ptr) 
compiler: ccache gcc -fPIC -pthread -m64 -Wall -O3 -g -DOPENSSL_USE_NODELETE -DL_ENDIAN -DOPENSSL_PIC -DNDEBUG
                              op      op/s
 521 bits ecdh (nistp521)   0.0038s    262.2
```

## After this PR (ca65f5b373c5ed29b132fa1aa9a0ebbc1e874807)

```
> openssl speed ecdhb571 ecdhbrp512r1 ecdhp521 ecdhp256
version: 3.0.0-dev
built on: built on: Fri Feb 15 19:17:59 2019 UTC
options:bn(64,64) rc4(16x,int) des(int) aes(partial) idea(int) blowfish(ptr) 
compiler: ccache gcc -fPIC -pthread -m64 -Wa,--noexecstack -Wall -O3 -g -DOPENSSL_USE_NODELETE -DL_ENDIAN -DOPENSSL_PIC -DOPENSSL_CPUID_OBJ -DOPENSSL_IA32_SSE2 -DOPENSSL_BN_ASM_MONT -DOPENSSL_BN_ASM_MONT5 -DOPENSSL_BN_ASM_GF2m -DSHA1_ASM -DSHA256_ASM -DSHA512_ASM -DKECCAK1600_ASM -DRC4_ASM -DMD5_ASM -DAES_ASM -DVPAES_ASM -DBSAES_ASM -DGHASH_ASM -DECP_NISTZ256_ASM -DX25519_ASM -DPADLOCK_ASM -DPOLY1305_ASM -DNDEBUG
                              op      op/s
 256 bits ecdh (nistp256)   0.0001s  19684.2
 521 bits ecdh (nistp521)   0.0003s   2961.1
 571 bits ecdh (nistb571)   0.0019s    530.9
 512 bits ecdh (brainpoolP512r1)   0.0014s    716.1

> openssl speed ecdhp521 # no-asm build
version: 3.0.0-dev
built on: built on: Fri Feb 15 19:43:40 2019 UTC
options:bn(64,64) rc4(int) des(int) aes(partial) idea(int) blowfish(ptr) 
compiler: ccache gcc -fPIC -pthread -m64 -Wall -O3 -g -DOPENSSL_USE_NODELETE -DL_ENDIAN -DOPENSSL_PIC -DNDEBUG
                              op      op/s
 521 bits ecdh (nistp521)   0.0039s    259.5
```