@romen 2019-02-15
c4d2e27 adds a `CHANGES` entry.

Of course doing this will prevent clean cherry-pick to 1.1.1 and 1.1.0.

Once this PR gets approved I will open separate backporting PRs for 1.1.1 and 1.1.0.