@mspncp 2019-02-17
> [c4d2e27](https://github.com/openssl/openssl/commit/c4d2e2779bc40862670a436e8eac2aaecfe11256) adds a `CHANGES` entry.

> Of course doing this will prevent clean cherry-pick to 1.1.1 and 1.1.0.

Technically, the CHANGES file won't cherry-pick without merge conflict. But it's one which can trivially resolved by a human: Simply discard the incoming changes (`git checkout --ours CHANGES`) then edit the CHANGES file and re-add your entry at the correct location. Finally,  `git add CHANGES ; git cherry-pick --continue`.  So IMHO no need to open a separate pull request for it.