@romen 2019-02-17
> So IMHO no need to open a separate pull request for it. (_Edit:_ If it's the only merge conflict.)

Yes, I could trivially do that edit for the CHANGES file during the cherry-pick, but what is the policy about it? I tend to err on the over-conservative interpretation when in doubt.