@mspncp 2019-02-17
I've done it previously, and I know that @levitte does it sometimes, too. But AFAIK there is no official policy about that.