@romen 2019-02-17
Merged into master.

PR for 1.1.1 and 1.1.0 will show up soon as in both cases, in addition to conflicts on `CHANGES`, `test/build.info` requires manual intervention during cherry-pick.

Thanks all!